% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_fast.R
\name{aggregate_fast}
\alias{aggregate_fast}
\title{Tiled, Out-of-Core Time Series Aggregation (Internal)}
\usage{
aggregate_fast(
  paths,
  index,
  output_names,
  user_region,
  tile_degrees,
  output_dir,
  verbose,
  aggregation_type = "mean"
)
}
\description{
Tiled, Out-of-Core Time Series Aggregation (Internal)
}
\keyword{internal}
