% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_terra.R
\name{aggregate_terra}
\alias{aggregate_terra}
\title{In-Memory Time Series Aggregation (Internal)}
\usage{
aggregate_terra(
  x,
  index,
  output_names,
  output_dir,
  gdal_opt,
  overwrite,
  verbose,
  aggregation_type = "mean"
)
}
\description{
In-Memory Time Series Aggregation (Internal)
}
\keyword{internal}
