% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_fast.R
\name{average_fast}
\alias{average_fast}
\title{Tiled, Out-of-Core Time Series Averaging (Internal)}
\usage{
average_fast(paths, index, output_names, user_region, tile_degrees, output_dir)
}
\description{
Tiled, Out-of-Core Time Series Averaging (Internal)
}
\keyword{internal}
