% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_terra.R
\name{average_terra}
\alias{average_terra}
\title{In-Memory Time Series Averaging (Internal)}
\usage{
average_terra(x, index, output_names, output_dir, gdal_opt, overwrite)
}
\description{
In-Memory Time Series Averaging (Internal)
}
\keyword{internal}
