% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bios_terra.R
\name{bio03_terra}
\alias{bio03_terra}
\title{bio03_terra: Isothermality}
\usage{
bio03_terra(bio02, bio07)
}
\arguments{
\item{bio02}{A single-layer `SpatRaster` of Mean Diurnal Range (Bio02).}

\item{bio07}{A single-layer `SpatRaster` of Temperature Range (Bio07).}
}
\value{
A single-layer `SpatRaster` with the calculated isothermality, named "bio03".
}
\description{
Calculates Isothermality, defined as (bio02 / bio07) * 100.
}
\keyword{internal}
