% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bios_fast.R
\name{bio04_fast}
\alias{bio04_fast}
\title{bio04_fast: Temperature Seasonality (Std Dev * 100)}
\usage{
bio04_fast(tavg, cell)
}
\arguments{
\item{tavg}{A numeric **matrix** of average temperatures. **Rows** represent spatial units (cells) 
and **columns** represent temporal units.}

\item{cell}{A vector of original cell IDs. Its length must be exactly equal to the number of rows in `tavg`.}
}
\value{
A **matrix** with dimensions `c(N, 2)`, where N is the number of input cells. 
  The columns are named "bio04" (seasonality) and "cell".
}
\description{
Calculates Temperature Seasonality, defined as the standard deviation of average temperatures across all temporal units (e.g., 12 months), multiplied by 100.
}
\keyword{internal}
