% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help2.R
\name{bio07_fun}
\alias{bio07_fun}
\title{bio07: Temperature Annual Range (bio05 - bio06)}
\usage{
bio07_fun(bio05V, bio06V, cell)
}
\arguments{
\item{bio05V}{Vector or single-column matrix of bio05 values.}

\item{bio06V}{Vector or single-column matrix of bio06 values.}

\item{cell}{Vector of original cell IDs.}
}
\value{
Matrix with "bio07", "cell".
}
\description{
Calculates the difference between bio05 and bio06.
}
\keyword{internal}
