% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help2.R
\name{bio08_fun}
\alias{bio08_fun}
\title{bio08: Mean Temperature of Wettest Period}
\usage{
bio08_fun(tperiod, pperiod_max_idx, period_length, cell)
}
\arguments{
\item{tperiod}{Matrix of temperature period sums (output from `var_periods`).}

\item{pperiod_max_idx}{Vector indicating the index (1-based) of the wettest period for each row.}

\item{period_length}{Integer. Number of units per period.}

\item{cell}{Vector of original cell IDs.}
}
\value{
Matrix with "bio08", "cell".
}
\description{
Calculates mean temperature of the period with the highest precipitation sum.
}
\keyword{internal}
