% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bios_terra.R
\name{bio08_terra}
\alias{bio08_terra}
\title{bio08_terra: Mean Temperature of Wettest Period}
\usage{
bio08_terra(tmp, wettest_period)
}
\arguments{
\item{tmp}{A `SpatRaster` object where each layer represents the mean temperature for a rolling period.}

\item{wettest_period}{A single-layer `SpatRaster` where cell values are integers indicating the layer index (1-based) of the wettest period.}
}
\value{
A single-layer `SpatRaster` containing the mean temperature of the wettest period, named "bio08".
}
\description{
Calculates the mean temperature of the specific rolling period identified as the wettest.
}
\keyword{internal}
