% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bios_terra.R
\name{bio09_terra}
\alias{bio09_terra}
\title{bio09_terra: Mean Temperature of Driest Period}
\usage{
bio09_terra(tmp, driest_period)
}
\arguments{
\item{tmp}{A `SpatRaster` object where each layer represents the mean temperature for a rolling period.}

\item{driest_period}{A single-layer `SpatRaster` where cell values are integers indicating the layer index (1-based) of the driest period.}
}
\value{
A single-layer `SpatRaster` containing the mean temperature of the driest period, named "bio09".
}
\description{
Calculates the mean temperature of the specific rolling period identified as the driest.
}
\keyword{internal}
