% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bios_fast.R
\name{bio10_fast}
\alias{bio10_fast}
\title{bio10_fast: Mean Temperature of Warmest Period}
\usage{
bio10_fast(tperiod, tperiod_max_idx, period_length, cell)
}
\arguments{
\item{tperiod}{A numeric **matrix** of temperature values (means) for each rolling period. 
**Rows** represent spatial units (cells). **Columns** represent the rolling periods.}

\item{tperiod_max_idx}{An integer **vector** indicating the column index (1-based) of the warmest period for each row. 
Its length must be exactly equal to the number of rows in `tperiod`.}

\item{period_length}{A single **integer** representing the number of units per period.}

\item{cell}{A vector of original cell IDs. Its length must be exactly equal to the number of rows in `tperiod`.}
}
\value{
A **matrix** with dimensions `c(N, 2)`, where N is the number of input cells. 
  The columns are named "bio10" (mean temperature of warmest period) and "cell".
}
\description{
Calculates the mean temperature of the specific rolling period identified as the warmest (highest temperature).
}
\keyword{internal}
