% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bios_terra.R
\name{bio10_terra}
\alias{bio10_terra}
\title{bio10_terra: Mean Temperature of Warmest Period}
\usage{
bio10_terra(tmp, warmest_period)
}
\arguments{
\item{tmp}{A `SpatRaster` object where each layer represents the mean temperature for a rolling period.}

\item{warmest_period}{A single-layer `SpatRaster` where cell values are integers indicating the layer index (1-based) of the warmest period.}
}
\value{
A single-layer `SpatRaster` containing the mean temperature of the warmest period, named "bio10".
}
\description{
Calculates the mean temperature of the specific rolling period identified as the warmest.
}
\keyword{internal}
