% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bios_fast.R
\name{bio11_fast}
\alias{bio11_fast}
\title{bio11_fast: Mean Temperature of Coldest Period}
\usage{
bio11_fast(tperiod, tperiod_min_idx, period_length, cell)
}
\arguments{
\item{tperiod}{Matrix of temperature period sums.}

\item{tperiod_min_idx}{Vector indicating the index (1-based) of the coldest period.}

\item{period_length}{Integer. Number of units per period.}

\item{cell}{Vector of original cell IDs.}
}
\value{
Matrix with "bio11", "cell".
}
\description{
Calculates mean temperature of the period with the lowest temperature sum.
}
\keyword{internal}
