% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bios_terra.R
\name{bio11_terra}
\alias{bio11_terra}
\title{bio11_terra: Mean Temperature of Coldest Period}
\usage{
bio11_terra(tmp, coldest_period)
}
\arguments{
\item{tmp}{A `SpatRaster` object where each layer represents the mean temperature for a rolling period.}

\item{coldest_period}{A single-layer `SpatRaster` where cell values are integers indicating the layer index (1-based) of the coldest period.}
}
\value{
A single-layer `SpatRaster` containing the mean temperature of the coldest period, named "bio11".
}
\description{
Calculates the mean temperature of the specific rolling period identified as the coldest.
}
\keyword{internal}
