% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bios_fast.R
\name{bio12_fast}
\alias{bio12_fast}
\title{bio12_fast: Total Precipitation}
\usage{
bio12_fast(prcp, cell)
}
\arguments{
\item{prcp}{A numeric **matrix** of precipitation values. **Rows** represent spatial units (cells) 
and **columns** represent temporal units (e.g., months).}

\item{cell}{A vector of original cell IDs. Its length must be exactly equal to the number of rows in `prcp`.}
}
\value{
A **matrix** with dimensions `c(N, 2)`, where N is the number of input cells. 
  The columns are named "bio12" (total precipitation) and "cell".
}
\description{
Calculates the total precipitation (sum) across all temporal units (e.g., 12 months).
}
\keyword{internal}
