% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help2.R
\name{bio12_fun}
\alias{bio12_fun}
\title{bio12: Total Precipitation}
\usage{
bio12_fun(precp, cell)
}
\arguments{
\item{precp}{Matrix of precipitation values for each unit.}

\item{cell}{Vector of original cell IDs.}
}
\value{
Matrix with "bio12", "cell".
}
\description{
Calculates the sum of precipitation values across all units.
}
\keyword{internal}
