% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bios_terra.R
\name{bio12_terra}
\alias{bio12_terra}
\title{bio12_terra: Total Precipitation}
\usage{
bio12_terra(prcp)
}
\arguments{
\item{prcp}{A `SpatRaster` object where each layer represents precipitation for a temporal unit.}
}
\value{
A single-layer `SpatRaster` with the total precipitation, named "bio12".
}
\description{
Calculates the total precipitation (sum) across all temporal units (layers).
}
\keyword{internal}
