% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bios_terra.R
\name{bio13_terra}
\alias{bio13_terra}
\title{bio13_terra: Precipitation of Wettest Unit}
\usage{
bio13_terra(prcp, wettest_unit = NULL)
}
\arguments{
\item{prcp}{A `SpatRaster` object where each layer represents precipitation for a temporal unit.}

\item{wettest_unit}{(Optional) A single-layer `SpatRaster` where cell values are integers indicating a static layer index (1-based) from which to extract the value. If `NULL` (the default), the overall maximum across all layers is calculated.}
}
\value{
A single-layer `SpatRaster` with the precipitation of the wettest unit, named "bio13".
}
\description{
Identifies the precipitation of the wettest temporal unit (layer).
}
\keyword{internal}
