% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bios_terra.R
\name{bio14_terra}
\alias{bio14_terra}
\title{bio14_terra: Precipitation of Driest Unit}
\usage{
bio14_terra(prcp, driest_unit = NULL)
}
\arguments{
\item{prcp}{A `SpatRaster` object where each layer represents precipitation for a temporal unit.}

\item{driest_unit}{(Optional) A single-layer `SpatRaster` where cell values are integers indicating a static layer index (1-based) from which to extract the value. If `NULL` (the default), the overall minimum across all layers is calculated.}
}
\value{
A single-layer `SpatRaster` with the precipitation of the driest unit, named "bio14".
}
\description{
Identifies the precipitation of the driest temporal unit (layer).
}
\keyword{internal}
