% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bios_terra.R
\name{bio15_terra}
\alias{bio15_terra}
\title{bio15_terra: Precipitation Seasonality (CV)}
\usage{
bio15_terra(prcp)
}
\arguments{
\item{prcp}{A `SpatRaster` object where each layer represents precipitation for a temporal unit.}
}
\value{
A single-layer `SpatRaster` with the precipitation seasonality, named "bio15".
}
\description{
Calculates the Coefficient of Variation (CV) of precipitation across all layers.
}
\note{
The formula adds 1 to the mean to avoid division by zero in arid areas.
}
\keyword{internal}
