% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bios_fast.R
\name{bio16_fast}
\alias{bio16_fast}
\title{bio16_fast: Precipitation of Wettest Period}
\usage{
bio16_fast(pperiod, pperiod_max_idx, cell)
}
\arguments{
\item{pperiod}{A numeric **matrix** of precipitation sums for each rolling period. 
**Rows** represent spatial units (cells). **Columns** represent the rolling periods.}

\item{pperiod_max_idx}{An integer **vector** indicating the column index (1-based) of the wettest period for each row. 
Its length must be exactly equal to the number of rows in `pperiod`.}

\item{cell}{A vector of original cell IDs. Its length must be exactly equal to the number of rows in `pperiod`.}
}
\value{
A **matrix** with dimensions `c(N, 2)`, where N is the number of input cells.
  The columns are named "bio16" (precipitation of wettest period) and "cell".
}
\description{
Calculates the total precipitation of the specific rolling period identified as the wettest (highest precipitation).
}
\keyword{internal}
