% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bios_terra.R
\name{bio16_terra}
\alias{bio16_terra}
\title{bio16_terra: Precipitation of Wettest Period}
\usage{
bio16_terra(wet, wettest_period)
}
\arguments{
\item{wet}{A `SpatRaster` object where each layer is the precipitation sum for a rolling period.}

\item{wettest_period}{A single-layer `SpatRaster` where cell values are integers indicating the layer index (1-based) of the wettest period.}
}
\value{
A single-layer `SpatRaster` with the precipitation of the wettest period, named "bio16".
}
\description{
Calculates the total precipitation of the specific rolling period identified as the wettest.
}
\keyword{internal}
