% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bios_fast.R
\name{bio17_fast}
\alias{bio17_fast}
\title{bio17_fast: Precipitation of Driest Period}
\usage{
bio17_fast(pperiod, pperiod_min_idx, cell)
}
\arguments{
\item{pperiod}{A numeric **matrix** of precipitation sums for each rolling period. 
**Rows** represent spatial units (cells). **Columns** represent the rolling periods.}

\item{pperiod_min_idx}{An integer **vector** indicating the column index (1-based) of the driest period for each row. 
Its length must be exactly equal to the number of rows in `pperiod`.}

\item{cell}{A vector of original cell IDs. Its length must be exactly equal to the number of rows in `pperiod`.}
}
\value{
A **matrix** with dimensions `c(N, 2)`, where N is the number of input cells.
  The columns are named "bio17" (precipitation of driest period) and "cell".
}
\description{
Calculates the total precipitation of the specific rolling period identified as the driest (lowest precipitation).
}
\keyword{internal}
