% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bios_terra.R
\name{bio17_terra}
\alias{bio17_terra}
\title{bio17_terra: Precipitation of Driest Period}
\usage{
bio17_terra(wet, driest_period)
}
\arguments{
\item{wet}{A `SpatRaster` object where each layer is the precipitation sum for a rolling period.}

\item{driest_period}{A single-layer `SpatRaster` where cell values are integers indicating the layer index (1-based) of the driest period.}
}
\value{
A single-layer `SpatRaster` with the precipitation of the driest period, named "bio17".
}
\description{
Calculates the total precipitation of the specific rolling period identified as the driest.
}
\keyword{internal}
