% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help2.R
\name{bio18_fun}
\alias{bio18_fun}
\title{BIO18: Precipitation of Warmest Period}
\usage{
bio18_fun(pperiod, tperiod_max_idx, cell)
}
\arguments{
\item{pperiod}{Matrix of precipitation period sums.}

\item{tperiod_max_idx}{Vector indicating the index (1-based) of the warmest period.}

\item{cell}{Vector of original cell IDs.}
}
\value{
Matrix with "bio18", "cell".
}
\description{
Calculates precipitation sum of the period with the highest temperature sum.
}
\keyword{internal}
