% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bios_terra.R
\name{bio18_terra}
\alias{bio18_terra}
\title{bio18_terra: Precipitation of Warmest Period}
\usage{
bio18_terra(wet, warmest_period)
}
\arguments{
\item{wet}{A `SpatRaster` object where each layer is the precipitation sum for a rolling period.}

\item{warmest_period}{A single-layer `SpatRaster` where cell values are integers indicating the layer index (1-based) of the warmest period.}
}
\value{
A single-layer `SpatRaster` with the precipitation of the warmest period, named "bio18".
}
\description{
Calculates the total precipitation of the specific rolling period identified as the warmest.
}
\keyword{internal}
