% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bios_fast.R
\name{bio19_fast}
\alias{bio19_fast}
\title{bio19_fast: Precipitation of Coldest Period}
\usage{
bio19_fast(pperiod, tperiod_min_idx, cell)
}
\arguments{
\item{pperiod}{A numeric **matrix** of precipitation sums for each rolling period. 
**Rows** represent spatial units (cells). **Columns** represent the rolling periods.}

\item{tperiod_min_idx}{An integer **vector** indicating the column index (1-based) of the coldest period for each row. 
Its length must be exactly equal to the number of rows in `pperiod`.}

\item{cell}{A vector of original cell IDs. Its length must be exactly equal to the number of rows in `pperiod`.}
}
\value{
A **matrix** with dimensions `c(N, 2)`, where N is the number of input cells.
  The columns are named "bio19" (precipitation of coldest period) and "cell".
}
\description{
Calculates the total precipitation of the specific rolling period identified as the coldest (lowest temperature).
}
\keyword{internal}
