% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bios_terra.R
\name{bio19_terra}
\alias{bio19_terra}
\title{bio19_terra: Precipitation of Coldest Period}
\usage{
bio19_terra(wet, coldest_period)
}
\arguments{
\item{wet}{A `SpatRaster` object where each layer is the precipitation sum for a rolling period.}

\item{coldest_period}{A single-layer `SpatRaster` where cell values are integers indicating the layer index (1-based) of the coldest period.}
}
\value{
A single-layer `SpatRaster` with the precipitation of the coldest period, named "bio19".
}
\description{
Calculates the total precipitation of the specific rolling period identified as the coldest.
}
\keyword{internal}
