% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help2.R
\name{bio1_fun}
\alias{bio1_fun}
\title{BIO1: Mean Temperature of Units}
\usage{
bio1_fun(tavg, cell)
}
\arguments{
\item{tavg}{Matrix of average temperatures for each unit.}

\item{cell}{Vector of original cell IDs.}
}
\value{
Matrix with columns: "bio1", "cell".
}
\description{
Calculates mean temperature across all temporal units.
}
\keyword{internal}
