% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bios_terra.R
\name{bio20_terra}
\alias{bio20_terra}
\title{bio20_terra: Mean Radiation}
\usage{
bio20_terra(srad)
}
\arguments{
\item{srad}{A `SpatRaster` object where each layer represents solar radiation for a temporal unit.}
}
\value{
A single-layer `SpatRaster` with the mean solar radiation, named "bio20".
}
\description{
Calculates mean solar radiation across all temporal units (layers).
}
\keyword{internal}
