% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bios_terra.R
\name{bio22_terra}
\alias{bio22_terra}
\title{bio22_terra: Lowest Radiation Unit}
\usage{
bio22_terra(srad, low_rad_unit = NULL)
}
\arguments{
\item{srad}{A `SpatRaster` object where each layer represents solar radiation for a temporal unit.}

\item{low_rad_unit}{(Optional) A single-layer `SpatRaster` where cell values are integers indicating a static layer index (1-based). If `NULL`, the overall minimum across all layers is calculated.}
}
\value{
A single-layer `SpatRaster` with the minimum solar radiation, named "bio22".
}
\description{
Identifies the lowest solar radiation of the unit with the lowest value.
}
\keyword{internal}
