% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bios_terra.R
\name{bio23_terra}
\alias{bio23_terra}
\title{bio23_terra: Radiation Seasonality (CV)}
\usage{
bio23_terra(srad)
}
\arguments{
\item{srad}{A `SpatRaster` object where each layer represents solar radiation for a temporal unit.}
}
\value{
A single-layer `SpatRaster` with the solar radiation seasonality, named "bio23".
}
\description{
Calculates the Coefficient of Variation (CV) of solar radiation across all layers.
}
\keyword{internal}
