% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help2.R
\name{bio25_fun}
\alias{bio25_fun}
\title{bio25: Solar Radiation of Driest Period}
\usage{
bio25_fun(speriod, pperiod_min_idx, cell)
}
\arguments{
\item{speriod}{Matrix of solar radiation period means (output from `var_periods`).}

\item{pperiod_min_idx}{Vector indicating the index (1-based) of the driest period.}

\item{cell}{Vector of original cell IDs.}
}
\value{
Matrix with "bio25", "cell".
}
\description{
Calculates solar radiation mean of the period with the highest precipitation sum.
}
\keyword{internal}
