% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help2.R
\name{bio26_fun}
\alias{bio26_fun}
\title{bio26: Solar Radiation of Warmest Period}
\usage{
bio26_fun(speriod, tperiod_max_idx, cell)
}
\arguments{
\item{speriod}{Matrix of solar radiation period means (output from `var_periods`).}

\item{tperiod_max_idx}{Vector indicating the index (1-based) of the warmest period.}

\item{cell}{Vector of original cell IDs.}
}
\value{
Matrix with "bio26", "cell".
}
\description{
Calculates solar radiation mean of the period with the highest temperature mean.
}
\keyword{internal}
