% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bios_terra.R
\name{bio26_terra}
\alias{bio26_terra}
\title{bio26_terra: Radiation of Warmest Period}
\usage{
bio26_terra(prad, warmest_period)
}
\arguments{
\item{prad}{A `SpatRaster` object where each layer is the mean solar radiation for a rolling period.}

\item{warmest_period}{A single-layer `SpatRaster` where cell values are integers indicating the layer index (1-based) of the warmest period.}
}
\value{
A single-layer `SpatRaster` with the solar radiation of the warmest period, named "bio26".
}
\description{
Calculates the mean solar radiation of the specific rolling period identified as the warmest.
}
\keyword{internal}
