% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bios_terra.R
\name{bio27_terra}
\alias{bio27_terra}
\title{bio27_terra: Radiation of Coldest Period}
\usage{
bio27_terra(prad, coldest_period)
}
\arguments{
\item{prad}{A `SpatRaster` object where each layer is the mean solar radiation for a rolling period.}

\item{coldest_period}{A single-layer `SpatRaster` where cell values are integers indicating the layer index (1-based) of the coldest period.}
}
\value{
A single-layer `SpatRaster` with the solar radiation of the coldest period, named "bio27".
}
\description{
Calculates the mean solar radiation of the specific rolling period identified as the coldest.
}
\keyword{internal}
