% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bios_terra.R
\name{bio29_terra}
\alias{bio29_terra}
\title{bio29_terra: Highest Moisture Unit}
\usage{
bio29_terra(mois, high_mois_unit = NULL)
}
\arguments{
\item{mois}{A `SpatRaster` object where each layer represents moisture for a temporal unit.}

\item{high_mois_unit}{(Optional) A single-layer `SpatRaster` where cell values are integers indicating a static layer index (1-based). If `NULL`, the overall maximum across all layers is calculated.}
}
\value{
A single-layer `SpatRaster` with the maximum moisture, named "bio29".
}
\description{
Identifies the highest moisture of the unit with the highest value.
}
\keyword{internal}
