% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help2.R
\name{bio2_fun}
\alias{bio2_fun}
\title{BIO2: Mean Diurnal Range}
\usage{
bio2_fun(tmin, tmax, cell)
}
\arguments{
\item{tmin}{Matrix of minimum temperatures for each unit.}

\item{tmax}{Matrix of maximum temperatures for each unit.}

\item{cell}{Vector of original cell IDs.}
}
\value{
Matrix with "bio2", "cell".
}
\description{
Calculates the mean of (tmax - tmin) across all temporal units.
}
\keyword{internal}
