% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bios_terra.R
\name{bio30_terra}
\alias{bio30_terra}
\title{bio30_terra: Lowest Moisture Unit}
\usage{
bio30_terra(mois, low_mois_unit = NULL)
}
\arguments{
\item{mois}{A `SpatRaster` object where each layer represents moisture for a temporal unit.}

\item{low_mois_unit}{(Optional) A single-layer `SpatRaster` where cell values are integers indicating a static layer index (1-based). If `NULL`, the overall minimum across all layers is calculated.}
}
\value{
A single-layer `SpatRaster` with the minimum moisture, named "bio30".
}
\description{
Identifies the lowest moisture of the unit with the lowest value.
}
\keyword{internal}
