% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help2.R
\name{bio31_fun}
\alias{bio31_fun}
\title{bio31: Moisture Seasonality (Standard Deviation)}
\usage{
bio31_fun(mois, n_units, cell)
}
\arguments{
\item{mois}{Matrix containing moisture values for each unit.}

\item{n_units}{Integer. The total number of temporal units.}

\item{cell}{Vector of original cell IDs.}
}
\value{
Matrix with "bio31", "cell".
}
\description{
Calculates coefficient of variation in moisture across units.
}
\keyword{internal}
