% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bios_terra.R
\name{bio31_terra}
\alias{bio31_terra}
\title{bio31_terra: Moisture Seasonality (Std Dev * 100)}
\usage{
bio31_terra(mois)
}
\arguments{
\item{mois}{A `SpatRaster` object where each layer represents moisture for a temporal unit.}
}
\value{
A single-layer `SpatRaster` with the moisture seasonality, named "bio31".
}
\description{
Calculates the standard deviation of moisture across all layers, multiplied by 100.
}
\keyword{internal}
