% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bios_fast.R
\name{bio32_fast}
\alias{bio32_fast}
\title{bio32_fast: Mean Moisture of Most Moist Period}
\usage{
bio32_fast(speriod, speriod_max_idx, cell)
}
\arguments{
\item{speriod}{A numeric **matrix** of moisture values (means) for each rolling period. 
**Rows** represent spatial units (cells). **Columns** represent the rolling periods.}

\item{speriod_max_idx}{An integer **vector** indicating the column index (1-based) of the most moist period for each row. 
Its length must be exactly equal to the number of rows in `speriod`.}

\item{cell}{A vector of original cell IDs. Its length must be exactly equal to the number of rows in `speriod`.}
}
\value{
A **matrix** with dimensions `c(N, 2)`, where N is the number of input cells.
  The columns are named "bio32" (mean moisture of most moist period) and "cell".
}
\description{
Calculates the mean moisture of the specific rolling period identified as the most moist (highest moisture).
}
\keyword{internal}
