% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bios_terra.R
\name{bio32_terra}
\alias{bio32_terra}
\title{bio32_terra: Mean Moisture of Most Moist Period}
\usage{
bio32_terra(pmois, high_mois_period)
}
\arguments{
\item{pmois}{A `SpatRaster` object where each layer is the mean moisture for a rolling period.}

\item{high_mois_period}{A single-layer `SpatRaster` where cell values are integers indicating the layer index (1-based) of the most moist period.}
}
\value{
A single-layer `SpatRaster` with the moisture of the most moist period, named "bio32".
}
\description{
Calculates the mean moisture of the specific rolling period identified as the most moist.
}
\keyword{internal}
