% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help2.R
\name{bio33_fun}
\alias{bio33_fun}
\title{bio33: Moisture of the Least Moist Period}
\usage{
bio33_fun(speriod, speriod_min_idx, cell)
}
\arguments{
\item{speriod}{Matrix of moisture period means (output from `var_periods`).}

\item{speriod_min_idx}{Vector indicating the index (1-based) of the least moist period.}

\item{cell}{Vector of original cell IDs.}
}
\value{
Matrix with "bio33", "cell".
}
\description{
Calculates moisture mean of the least moist period.
}
\keyword{internal}
