% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bios_fast.R
\name{bio34_fast}
\alias{bio34_fast}
\title{bio34_fast: Moisture of Warmest Period}
\usage{
bio34_fast(speriod, tperiod_max_idx, cell)
}
\arguments{
\item{speriod}{Matrix of moisture period means (output from `var_periods`).}

\item{tperiod_max_idx}{Vector indicating the index (1-based) of the warmest period.}

\item{cell}{Vector of original cell IDs.}
}
\value{
Matrix with "bio34", "cell".
}
\description{
Calculates moisture mean of the period with the highest temperature mean.
}
\keyword{internal}
