% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bios_terra.R
\name{bio34_terra}
\alias{bio34_terra}
\title{bio34_terra: Mean Moisture of Warmest Period}
\usage{
bio34_terra(pmois, warmest_period)
}
\arguments{
\item{pmois}{A `SpatRaster` object where each layer is the mean moisture for a rolling period.}

\item{warmest_period}{A single-layer `SpatRaster` where cell values are integers indicating the layer index (1-based) of the warmest period.}
}
\value{
A single-layer `SpatRaster` with the moisture of the warmest period, named "bio34".
}
\description{
Calculates the mean moisture of the specific rolling period identified as the warmest.
}
\keyword{internal}
