% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help2.R
\name{bio3_fun}
\alias{bio3_fun}
\title{BIO3: Isothermality}
\usage{
bio3_fun(bio2V, bio7V, cell)
}
\arguments{
\item{bio2V}{Vector or single-column matrix of BIO2 values.}

\item{bio7V}{Vector or single-column matrix of BIO7 values.}

\item{cell}{Vector of original cell IDs.}
}
\value{
Matrix with "bio3", "cell".
}
\description{
Calculates (BIO2 / BIO7) * 100.
}
\keyword{internal}
