% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help2.R
\name{bio4_fun}
\alias{bio4_fun}
\title{BIO4: Temperature Seasonality (Std Dev * 100)}
\usage{
bio4_fun(tavg, cell)
}
\arguments{
\item{tavg}{Matrix of average temperatures for each unit.}

\item{cell}{Vector of original cell IDs.}
}
\value{
Matrix with "bio4", "cell".
}
\description{
Calculates the standard deviation of average temperatures across units, multiplied by 100.
}
\keyword{internal}
