% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help2.R
\name{bio7_fun}
\alias{bio7_fun}
\title{BIO7: Temperature Annual Range (BIO5 - BIO6)}
\usage{
bio7_fun(bio5V, bio6V, cell)
}
\arguments{
\item{bio5V}{Vector or single-column matrix of BIO5 values.}

\item{bio6V}{Vector or single-column matrix of BIO6 values.}

\item{cell}{Vector of original cell IDs.}
}
\value{
Matrix with "bio7", "cell".
}
\description{
Calculates the difference between BIO5 and BIO6.
}
\keyword{internal}
