% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help2.R
\name{bio9_fun}
\alias{bio9_fun}
\title{BIO9: Mean Temperature of Driest Period}
\usage{
bio9_fun(tperiod, pperiod_min_idx, period_length, cell)
}
\arguments{
\item{tperiod}{Matrix of temperature period sums.}

\item{pperiod_min_idx}{Vector indicating the index (1-based) of the driest period.}

\item{period_length}{Integer. Number of units per period.}

\item{cell}{Vector of original cell IDs.}
}
\value{
Matrix with "bio9", "cell".
}
\description{
Calculates mean temperature of the period with the lowest precipitation sum.
}
\keyword{internal}
