% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_01.R
\name{bio_01}
\alias{bio_01}
\title{bio_01  Create annual mean temperature (BIO1)}
\usage{
bio_01(tas, filename = "")
}
\arguments{
\item{tas}{A SpatRaster with 12 layers (monthly mean temperature).}

\item{filename}{Optional file path to write the output raster.}
}
\value{
A SpatRaster with one layer representing BIO1.
}
\description{
bio_01  Create annual mean temperature (BIO1)
}
\examples{
tas_example <- mock_tas()
bio_01(tas_example)
}
