% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_02.R
\name{bio_02}
\alias{bio_02}
\title{bio_02  Create mean diurnal temperature range}
\usage{
bio_02(tasmax, tasmin, filename = "")
}
\arguments{
\item{tasmax}{Maximum monthly temperature. A set of 12 raster of maximum temperature.}

\item{tasmin}{Minimum monthly temperature. A set of 12 raster of minimum temperature.}

\item{filename}{the output where to write}
}
\value{
a raster with the bio_02 variable
}
\description{
bio_02  Create mean diurnal temperature range
}
\examples{
tmax <- mock_tasmax()
tmin <- mock_tasmin()
bio_02(tmax, tmin)
}
