% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_1.R
\name{bio_1}
\alias{bio_1}
\title{bio_1 Create monthly average temperature}
\usage{
bio_1(tas, filename = "")
}
\arguments{
\item{tas}{Average monthly temperature raster with 12 layers}

\item{filename}{the output where to write}
}
\value{
a raster with the bio_1 variable
}
\description{
bio_1 Create monthly average temperature
}
