% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_10.R
\name{bio_10}
\alias{bio_10}
\title{bio 10. Mean Temperature of the warmest Quarter}
\usage{
bio_10(tas)
}
\arguments{
\item{tas}{}

\item{filename}{}
}
\description{
bio 10. Mean Temperature of the warmest Quarter
}
