% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_11.R
\name{bio_11}
\alias{bio_11}
\title{bio 11. Mean Temperature of the coldest Quarter}
\usage{
bio_11(tas)
}
\arguments{
\item{tas}{}

\item{filename}{}
}
\description{
bio 11. Mean Temperature of the coldest Quarter
}
