% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_12.R
\name{bio_12}
\alias{bio_12}
\title{bio_12 Annual Precipitation Sum}
\usage{
bio_12(pr, filename = "")
}
\arguments{
\item{pr}{Monthly precipitation raster with 12 layers}

\item{filename}{the output where to write}
}
\value{
a raster with the bio_12 variable
}
\description{
bio_12 Annual Precipitation Sum
}
