% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_12_m.R
\name{bio_12_m}
\alias{bio_12_m}
\title{bio_12 Annual precipitation sum. Matrix
version.}
\usage{
bio_12_m(pr)
}
\arguments{
\item{pr}{A matrix with 12 columns of monthly precipitation}
}
\value{
a vector with the bio_12 variable
}
\description{
bio_12 Annual precipitation sum. Matrix
version.
}
