% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_12_optim.R
\name{bio_12_optim}
\alias{bio_12_optim}
\title{Optimized BIO12 computation using RcppParallel}
\usage{
bio_12_optim(
  pr,
  filename = "",
  na_rm = FALSE,
  nthreads = parallel::detectCores(),
  block_factor = 4
)
}
\arguments{
\item{pr}{SpatRaster with 12 layers (monthly precipitation)}

\item{filename}{Optional output file path}

\item{na_rm}{Logical, remove NA values}
}
\value{
SpatRaster with one layer (annual precipitation)
}
\description{
Optimized BIO12 computation using RcppParallel
}
