% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_13.R
\name{bio_13}
\alias{bio_13}
\title{Bio 13. Precipitation of wettest month}
\usage{
bio_13(pr, filename = "")
}
\arguments{
\item{pr}{Yearly precipitation}

\item{filename}{The output filename}
}
\value{
A raster with the wettest month of the year
}
\description{
Bio 13. Precipitation of wettest month
}
