% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_13_m.R
\name{bio_13_m}
\alias{bio_13_m}
\title{bio_13 """Precipitation of wettest month. Matrix
version.}
\usage{
bio_13_m(pr)
}
\arguments{
\item{pr}{A matrix with 12 columns of monthly precipitation}
}
\value{
a vector with the bio_13 variable
}
\description{
bio_13 """Precipitation of wettest month. Matrix
version.
}
