% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_14.R
\name{bio_14}
\alias{bio_14}
\title{Bio 14. Precipitation of driest month}
\usage{
bio_14(pr, filename = "")
}
\arguments{
\item{pr}{Yearly precipitation}

\item{filename}{The output filename}
}
\value{
A raster with the driest month of the year
}
\description{
Bio 14. Precipitation of driest month
}
