% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_14_m.R
\name{bio_14_m}
\alias{bio_14_m}
\title{bio_14 Precipitation of driest month. Matrix
version.}
\usage{
bio_14_m(pr)
}
\arguments{
\item{pr}{A matrix with 12 columns of monthly precipitation}
}
\value{
a vector with the bio_14 variable
}
\description{
bio_14 Precipitation of driest month. Matrix
version.
}
