% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_15.R
\name{bio_15}
\alias{bio_15}
\title{bio_15 Coefficient of variation in precipitation rate}
\usage{
bio_15(pr, filename = "")
}
\arguments{
\item{pr}{Monthly precipitation raster with 12 layers}

\item{filename}{the output where to write}
}
\value{
a raster with the bio_15 variable
}
\description{
bio_15 Coefficient of variation in precipitation rate
}
