% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_15_m.R
\name{bio_15_m}
\alias{bio_15_m}
\title{bio_15 Precipitation Seasonality  
version.}
\usage{
bio_15_m(pr)
}
\arguments{
\item{pr}{A matrix with 12 columns of monthly precipitation}
}
\value{
a vector with the bio_15 variable
}
\description{
bio_15 Precipitation Seasonality  
version.
}
