% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_16_m.R
\name{bio_16_m}
\alias{bio_16_m}
\title{bio 16 Precipitation of Wettest Quarter}
\usage{
bio_16_m(pr)
}
\arguments{
\item{pr}{Monthly precipitation}
}
\value{
A raster with the precipitation of the Wettest Quarter
}
\description{
bio 16 Precipitation of Wettest Quarter
}
