% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_17_m.R
\name{bio_17_m}
\alias{bio_17_m}
\title{bio 17 Precipitation of driest Quarter}
\usage{
bio_17_m(pr)
}
\arguments{
\item{pr}{Monthly precipitation}
}
\value{
A matrix with the precipitation of the driest
}
\description{
bio 17 Precipitation of driest Quarter
}
