% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_19.R
\name{bio_19}
\alias{bio_19}
\title{bio 19, Precipitation of Coldest Quarter}
\usage{
bio_19(pr, tasmin, filename = "")
}
\arguments{
\item{pr}{Monthly precipiation}

\item{filename}{Output filename}

\item{tasmax}{Maximum monthly temperature}
}
\description{
bio 19, Precipitation of Coldest Quarter
}
