% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_1_m.R
\name{bio_1_m}
\alias{bio_1_m}
\title{bio_1 Create monthly average temperature. Matrix
version.}
\usage{
bio_1_m(tas)
}
\arguments{
\item{tas}{A matrix with 12 columns of the average monthly temperature.}
}
\value{
a vector with the bio_1 variable
}
\description{
bio_1 Create monthly average temperature. Matrix
version.
}
