% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_2.R
\name{bio_2}
\alias{bio_2}
\title{bio_1 Create monthly average temperature}
\usage{
bio_2(tasmax, tasmin, filename = "")
}
\arguments{
\item{tasmax}{Maximum monthly temperature. A set of 12 raster of maximum temperature.}

\item{tasmin}{Minimum monthly temperature. A set of 12 raster of minimum temperature.}

\item{filename}{the output where to write}
}
\value{
a raster with the bio_2 variable
}
\description{
bio_1 Create monthly average temperature
}
