% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_2_m.R
\name{bio_2_m}
\alias{bio_2_m}
\title{bio_2  Create mean diurnal temperature range. Matrix 
version.}
\usage{
bio_2_m(tasmax, tasmin)
}
\arguments{
\item{tasmax}{A matrix with 12 columns of the maximum monthly temperature.}

\item{tasmin}{A matrix with 12 columns of the maximum monthly temperature.}

\item{filename}{the output where to write}
}
\value{
a matrix with the bio_2 variable
}
\description{
bio_2  Create mean diurnal temperature range. Matrix 
version.
}
