% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_3_m.R
\name{bio_3_m}
\alias{bio_3_m}
\title{Bio 3. calculate bio3 "Isothermality"
The ratio of diurnal variation (bio2) to annual range of air temperature (bio7).
Is multiplied by a factor of 100}
\usage{
bio_3_m(tasmax, tasmin)
}
\arguments{
\item{tasmax}{Maximum monthly temperature.}

\item{tasmin}{Minimum monthly temperature.}
}
\value{
A vector with the range of temperatures
}
\description{
Bio 3. calculate bio3 "Isothermality"
The ratio of diurnal variation (bio2) to annual range of air temperature (bio7).
Is multiplied by a factor of 100
}
