% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_4.R
\name{bio_4}
\alias{bio_4}
\title{bio_4 Temperature seasonality}
\usage{
bio_4(tas, filename = "")
}
\arguments{
\item{tas}{Average monthly temperature raster with 12 layers}

\item{filename}{the output where to write}
}
\value{
a raster with the bio_5 variable. Standard deviation of the average monthly temperature
}
\description{
bio_4 Temperature seasonality
}
