% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_4_m.R
\name{bio_4_m}
\alias{bio_4_m}
\title{bio_1 Create standard deviation of the temperature. 
Matrix version.}
\usage{
bio_4_m(tas)
}
\arguments{
\item{tas}{A matrix with 12 columns of the average monthly temperature.}
}
\value{
a vector with the bio_4 variable
}
\description{
bio_1 Create standard deviation of the temperature. 
Matrix version.
}
