% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_5.R
\name{bio_5}
\alias{bio_5}
\title{Bio 5. Max Temperature of Warmest Month}
\usage{
bio_5(tasmax, filename = "")
}
\arguments{
\item{tasmax}{Maximum monthly temperature.}

\item{filename}{the output filename}
}
\value{
A raster with the maximum value of the monthly temperature time series
}
\description{
Bio 5. Max Temperature of Warmest Month
}
