% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_5_m.R
\name{bio_5_m}
\alias{bio_5_m}
\title{Bio 5. Max Temperature of Warmest Month}
\usage{
bio_5_m(tasmax)
}
\arguments{
\item{tasmax}{Maximum monthly temperature.}
}
\value{
A vector with the range of temperatures
}
\description{
Bio 5. Max Temperature of Warmest Month
}
