% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_6.R
\name{bio_6}
\alias{bio_6}
\title{Bio 5. Min Temperature of Coldest Month}
\usage{
bio_6(tasmin, filename = "")
}
\arguments{
\item{tasmin}{Minimum monthly temperature.}

\item{filename}{the output filename}
}
\value{
A raster with the minimum value of the monthly minimum temperature time series
}
\description{
Bio 5. Min Temperature of Coldest Month
}
