% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_6_m.R
\name{bio_6_m}
\alias{bio_6_m}
\title{Bio 6. Min Temperature of Coldest Month}
\usage{
bio_6_m(tasmin)
}
\arguments{
\item{tasmin}{Minimum monthly temperature.}
}
\value{
A vector with the minimum temperature  of coldest month
}
\description{
Bio 6. Min Temperature of Coldest Month
}
