% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_7.R
\name{bio_7}
\alias{bio_7}
\title{Bio 7. Max Temperature of Warmest Month}
\usage{
bio_7(tasmax, tasmin, filename = "")
}
\arguments{
\item{tasmax}{Maximum monthly temperature.}

\item{filename}{the output filename}
}
\value{
A raster with the maximum value of the monthly temperature time series
}
\description{
Bio 7. Max Temperature of Warmest Month
}
