% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_7_m.R
\name{bio_7_m}
\alias{bio_7_m}
\title{Bio 7. Isothermalithy. Difference between
Bio 5 and Bio 6}
\usage{
bio_7_m(tasmax, tasmin)
}
\arguments{
\item{tasmax}{Maximum monthly temperature.}

\item{tasmin}{Minimum monthly temperature.}

\item{filename}{the output filename}
}
\value{
A vector with the range of temperatures
}
\description{
Bio 7. Isothermalithy. Difference between
Bio 5 and Bio 6
}
