% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_9_m.R
\name{bio_9_m}
\alias{bio_9_m}
\title{bio 9. Mean Temperature of driest Quarter. First we
use a mobile window of three months to observer the 
quarter of the year (a three month slice) with lower
precipitation. After we get the index of this quarter
we average the temperature of the three months of 
this slice.}
\usage{
bio_9_m(pr, tas)
}
\arguments{
\item{pr}{Monthly precipitation}

\item{tas}{Monthly average temperature}
}
\value{
A raster with the mean temperature of the Wettest Quarter
}
\description{
bio 9. Mean Temperature of driest Quarter. First we
use a mobile window of three months to observer the 
quarter of the year (a three month slice) with lower
precipitation. After we get the index of this quarter
we average the temperature of the three months of 
this slice.
}
