% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chelsa_bio_12.R
\name{chelsa_bio_12}
\alias{chelsa_bio_12}
\title{Calculate bio_12 bioclimatic variable from Chelsa data}
\usage{
chelsa_bio_12(
  year,
  targetDir = "/media/reumanlab/backup/chelsav2_USA/",
  outputDir = "/media/reumanlab/backup/test_output/"
)
}
\arguments{
\item{year}{The target year to calculate the bioclim}

\item{targetDir}{The directory where chelsa variables are stored}

\item{outputDir}{The root output directory where to store the bioclimatic variables for that year}
}
\value{
A raster tif file with the bioclimatic variable 1
}
\description{
Calculate bio_12 bioclimatic variable from Chelsa data
}
