% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_average.R
\name{parallel_average}
\alias{parallel_average}
\title{Compute row means in parallel}
\usage{
parallel_average(mat)
}
\arguments{
\item{mat}{A numeric matrix}
}
\value{
A numeric matrix with one column containing row means
}
\description{
Compute row means in parallel
}
\examples{
mat <- matrix(1:9, nrow = 3)
parallel_average(mat)
}
