% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_difference.R
\name{parallel_difference}
\alias{parallel_difference}
\title{Compute element-wise difference in parallel}
\usage{
parallel_difference(mat_1, mat_2)
}
\arguments{
\item{mat_1}{A numeric matrix}

\item{mat_2}{A numeric matrix of the same dimensions as mat_1}
}
\value{
A numeric matrix with element-wise differences (mat_1 - mat_2)
}
\description{
Compute element-wise difference in parallel
}
\examples{
mat1 <- matrix(1:9, nrow = 3)
mat2 <- matrix(9:1, nrow = 3)
parallel_difference(mat1, mat2)
}
