% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcpp_parallel_average.R
\name{rcpp_parallel_average}
\alias{rcpp_parallel_average}
\title{Parallel row means of a numeric matrix}
\arguments{
\item{mat_1}{A numeric matrix. Each row will be averaged.}
}
\value{
A numeric matrix with one column, where each entry is the mean
  of the corresponding row in \code{mat_1}.
}
\description{
This function computes the mean of each row in a numeric matrix using
\pkg{RcppParallel}. It is designed to demonstrate parallel processing
with Rcpp and can be used for efficient row-wise averaging in large datasets.
}
\examples{
mat <- matrix(1:12, nrow = 3, byrow = TRUE)
rcpp_parallel_average(mat)
#      [,1]
# [1,]    2
# [2,]    6
# [3,]   10

}
\seealso{
\code{\link[RcppParallel]{parallelFor}}, \code{\link[Rcpp]{NumericMatrix}}
}
