% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll_terra.R
\name{roll_terra}
\alias{roll_terra}
\title{In-Memory Rolling Time Series Averaging (Internal)}
\usage{
roll_terra(
  x,
  window_size,
  freq,
  step,
  fun,
  output_names_list,
  output_dir,
  gdal_opt,
  overwrite
)
}
\description{
In-Memory Rolling Time Series Averaging (Internal)
}
\keyword{internal}
