% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bios_fast.R
\name{bio01_fast}
\alias{bio01_fast}
\title{bio01_fast: Mean Temperature of Units}
\usage{
bio01_fast(tavg, cell)
}
\arguments{
\item{tavg}{Matrix of average temperatures for each unit.}

\item{cell}{Vector of original cell IDs.}
}
\value{
Matrix with columns: "bio01", "cell".
}
\description{
Calculates mean temperature across all temporal units.
}
\keyword{internal}
