% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bios_terra.R
\name{bio01_terra}
\alias{bio01_terra}
\title{bio01_terra: Mean Temperature of Units}
\usage{
bio01_terra(tavg)
}
\arguments{
\item{tavg}{spatRaster of average temperatures for each unit.}
}
\value{
spatRaster with "bio01".
}
\description{
Calculates mean temperature across all temporal units.
}
\keyword{internal}
