% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bios_fast.R
\name{bio02_fast}
\alias{bio02_fast}
\title{bio02_fast: Mean Diurnal Range}
\usage{
bio02_fast(tmin, tmax, cell)
}
\arguments{
\item{tmin}{Matrix of minimum temperatures for each unit.}

\item{tmax}{Matrix of maximum temperatures for each unit.}

\item{cell}{Vector of original cell IDs.}
}
\value{
Matrix with "bio02", "cell".
}
\description{
Calculates the mean of (tmax - tmin) across all temporal units.
}
\keyword{internal}
