% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bios_terra.R
\name{bio02_terra}
\alias{bio02_terra}
\title{bio02_terra: Mean Diurnal Range}
\usage{
bio02_terra(tmin, tmax)
}
\arguments{
\item{tmin}{spatRaster of minimum temperatures for each unit.}

\item{tmax}{spatRaster of maximum temperatures for each unit.}
}
\value{
spatRaster with "bio02".
}
\description{
Calculates the mean of (tmax - tmin) across all temporal units.
}
\keyword{internal}
