% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bios_terra.R
\name{bio03_terra}
\alias{bio03_terra}
\title{bio03_terra: Isothermality}
\usage{
bio03_terra(bio02, bio07)
}
\arguments{
\item{bio02}{spatRaster of bio02 values.}

\item{bio07}{spatRaster of bio07 values.}
}
\value{
spatRaster with "bio03".
}
\description{
Calculates (bio02 / bio07) * 100.
}
\keyword{internal}
