% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bios_terra.R
\name{bio04_terra}
\alias{bio04_terra}
\title{bio04_terra: Temperature Seasonality (Std Dev * 100)}
\usage{
bio04_terra(tavg)
}
\arguments{
\item{tavg}{spatRaster of average temperatures for each unit.}
}
\value{
spatRaster with "bio04".
}
\description{
Calculates the standard deviation of average temperatures across units, multiplied by 100.
}
\keyword{internal}
