% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bios_terra.R
\name{bio06_terra}
\alias{bio06_terra}
\title{bio06_terra: Min Temperature of Coldest Unit}
\usage{
bio06_terra(tmin, coldest_unit = NULL)
}
\arguments{
\item{tmin}{spatRaster of minimum temperatures for each unit.}
}
\value{
spatRaster with "bio06".
}
\description{
Identifies min temperature of the coldest unit, potentially using a static index.
}
\keyword{internal}
