% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bios_terra.R
\name{bio07_terra}
\alias{bio07_terra}
\title{bio07_terra: Temperature Annual Range (bio05 - bio06)}
\usage{
bio07_terra(bio05, bio06)
}
\arguments{
\item{bio05}{spatRaster of bio05 values.}

\item{bio06}{spatRaster of bio06 values.}
}
\value{
spatRaster with "bio07".
}
\description{
Calculates the difference between bio05 and bio06.
}
\keyword{internal}
