% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bios_terra.R
\name{bio08_terra}
\alias{bio08_terra}
\title{bio08_terra: Mean Temperature of Wettest Period}
\usage{
bio08_terra(tmp, wettest_period)
}
\arguments{
\item{tmp}{spatRaster of temperature period sums.}

\item{wettest_period}{spatRaster indicating the index (1-based) of the wettest period for each cell.}
}
\value{
spatRaster with "bio08".
}
\description{
Calculates mean temperature of the period with the highest precipitation sum.
}
\keyword{internal}
