% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bios_terra.R
\name{bio09_terra}
\alias{bio09_terra}
\title{bio09_terra: Mean Temperature of Driest Period}
\usage{
bio09_terra(tmp, driest_period)
}
\arguments{
\item{tmp}{spatRaster of temperature period sums.}

\item{driest_period}{Vector indicating the index (1-based) of the driest period.}
}
\value{
spatRaster with "bio09".
}
\description{
Calculates mean temperature of the period with the lowest precipitation sum.
}
\keyword{internal}
