% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bios_fast.R
\name{bio10_fast}
\alias{bio10_fast}
\title{bio10_fast: Mean Temperature of Warmest Period}
\usage{
bio10_fast(tperiod, tperiod_max_idx, period_length, cell)
}
\arguments{
\item{tperiod}{Matrix of temperature period sums.}

\item{tperiod_max_idx}{Vector indicating the index (1-based) of the warmest period.}

\item{period_length}{Integer. Number of units per period.}

\item{cell}{Vector of original cell IDs.}
}
\value{
Matrix with "bio10", "cell".
}
\description{
Calculates mean temperature of the period with the highest temperature sum.
}
\keyword{internal}
