% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bios_terra.R
\name{bio10_terra}
\alias{bio10_terra}
\title{bio10_terra: Mean Temperature of Warmest Period}
\usage{
bio10_terra(tmp, warmest_period)
}
\arguments{
\item{tmp}{spatRaster of temperature period sums.}

\item{warmest_period}{Vector indicating the index (1-based) of the warmest period.}
}
\value{
spatRaster with "bio10".
}
\description{
Calculates mean temperature of the period with the highest temperature sum.
}
\keyword{internal}
