% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bios_terra.R
\name{bio11_terra}
\alias{bio11_terra}
\title{bio11_terra: Mean Temperature of Coldest Period}
\usage{
bio11_terra(tmp, coldest_period)
}
\arguments{
\item{tmp}{spatRaster of temperature period sums.}

\item{coldest_period}{Vector indicating the index (1-based) of the coldest period.}
}
\value{
spatRaster with "bio11".
}
\description{
Calculates mean temperature of the period with the lowest temperature sum.
}
\keyword{internal}
