% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bios_fast.R
\name{bio12_fast}
\alias{bio12_fast}
\title{bio12_fast: Total Precipitation}
\usage{
bio12_fast(precp, cell)
}
\arguments{
\item{precp}{Matrix of precipitation values for each unit.}

\item{cell}{Vector of original cell IDs.}
}
\value{
Matrix with "bio12", "cell".
}
\description{
Calculates the sum of precipitation values across all units.
}
\keyword{internal}
