% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bios_terra.R
\name{bio12_terra}
\alias{bio12_terra}
\title{bio12_terra: Total Precipitation}
\usage{
bio12_terra(prcp)
}
\arguments{
\item{prcp}{spatRaster of precipitation values for each unit.}
}
\value{
spatRaster with "bio12".
}
\description{
Calculates the sum of precipitation values across all units.
}
\keyword{internal}
