% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bios_terra.R
\name{bio13_terra}
\alias{bio13_terra}
\title{bio13_terra: Precipitation of Wettest Unit}
\usage{
bio13_terra(prcp, wettest_unit = NULL)
}
\arguments{
\item{prcp}{spatRaster of precipitation values for each unit.}
}
\value{
spatRaster with "bio13".
}
\description{
Identifies precipitation of the wettest unit, potentially using a static index.
}
\keyword{internal}
