% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bios_terra.R
\name{bio14_terra}
\alias{bio14_terra}
\title{bio14_terra: Precipitation of Driest Unit}
\usage{
bio14_terra(prcp, driest_unit = NULL)
}
\arguments{
\item{prcp}{spatRaster of precipitation values for each unit.}
}
\value{
spatRaster with "bio14".
}
\description{
Identifies precipitation of the driest unit, potentially using a static index.
}
\keyword{internal}
