% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bios_fast.R
\name{bio15_fast}
\alias{bio15_fast}
\title{bio15_fast: Precipitation Seasonality (CV)}
\usage{
bio15_fast(precp, bio12V, n_units, cell)
}
\arguments{
\item{precp}{Matrix containing precipitation values for each unit.}

\item{bio12V}{Precomputed total precipitation (BIO12 value).}

\item{n_units}{Integer. The total number of temporal units.}

\item{cell}{Vector of original cell IDs.}
}
\value{
Matrix with "bio15", "cell".
}
\description{
Calculates coefficient of variation in precipitation across units.
}
\keyword{internal}
