% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bios_terra.R
\name{bio15_terra}
\alias{bio15_terra}
\title{bio15_terra: Precipitation Seasonality (CV)}
\usage{
bio15_terra(prcp)
}
\arguments{
\item{prcp}{Matrix containing precipitation values for each unit.}
}
\value{
spatRaster with "bio15".
}
\description{
Calculates coefficient of variation in precipitation across units.
}
\note{
The "1 +" is to avoid strange CVs for areas where mean rainfaill is < 1)
}
\keyword{internal}
