% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bios_fast.R
\name{bio16_fast}
\alias{bio16_fast}
\title{bio16_fast: Precipitation of Wettest Period}
\usage{
bio16_fast(pperiod, pperiod_max_idx, cell)
}
\arguments{
\item{pperiod}{Matrix of precipitation period sums (output from `var_periods`).}

\item{pperiod_max_idx}{Vector indicating the index (1-based) of the wettest period.}

\item{cell}{Vector of original cell IDs.}
}
\value{
Matrix with "bio16", "cell".
}
\description{
Calculates precipitation sum of the period with the highest precipitation sum.
}
\keyword{internal}
