% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bios_fast.R
\name{bio17_fast}
\alias{bio17_fast}
\title{bio17_fast: Precipitation of Driest Period}
\usage{
bio17_fast(pperiod, pperiod_min_idx, cell)
}
\arguments{
\item{pperiod}{Matrix of precipitation period sums.}

\item{pperiod_min_idx}{Vector indicating the index (1-based) of the driest period.}

\item{cell}{Vector of original cell IDs.}
}
\value{
Matrix with "bio17", "cell".
}
\description{
Calculates precipitation sum of the period with the lowest precipitation sum.
}
\keyword{internal}
