% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bios_terra.R
\name{bio17_terra}
\alias{bio17_terra}
\title{bio17_terra: Precipitation of Driest Period}
\usage{
bio17_terra(wet, driest_period)
}
\arguments{
\item{wet}{spatRaster of precipitation period sums.}

\item{driest_period}{Vector indicating the index (1-based) of the driest period.}
}
\value{
spatRaster with "bio17".
}
\description{
Calculates precipitation sum of the period with the lowest precipitation sum.
}
\keyword{internal}
