% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bios_terra.R
\name{bio18_terra}
\alias{bio18_terra}
\title{bio18_terra: Precipitation of Warmest Period}
\usage{
bio18_terra(wet, warmest_period)
}
\arguments{
\item{wet}{spatRaster of precipitation period sums.}

\item{warmest_period}{Vector indicating the index (1-based) of the warmest period.}
}
\value{
spatRaster with "bio18".
}
\description{
Calculates precipitation sum of the period with the highest temperature sum.
}
\keyword{internal}
