% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bios_terra.R
\name{bio19_terra}
\alias{bio19_terra}
\title{bio19_terra: Precipitation of Coldest Period}
\usage{
bio19_terra(wet, coldest_period)
}
\arguments{
\item{wet}{spatRaster of precipitation period sums.}

\item{coldest_period}{Vector indicating the index (1-based) of the coldest period.}
}
\value{
spatRaster with "bio19".
}
\description{
Calculates precipitation sum of the period with the lowest temperature sum.
}
\keyword{internal}
