% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bios_terra.R
\name{bio20_terra}
\alias{bio20_terra}
\title{bio20_terra: Mean Solar Radiation of Units}
\usage{
bio20_terra(srad)
}
\arguments{
\item{srad}{spatRaster of average solar radiation for each unit.}
}
\value{
spatRaster with "bio20".
}
\description{
Calculates mean solar radiation across all temporal units.
}
\keyword{internal}
