% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bios_terra.R
\name{bio21_terra}
\alias{bio21_terra}
\title{bio21_terra: Highest Solar Radiation Unit}
\usage{
bio21_terra(srad, high_rad_unit = NULL)
}
\arguments{
\item{srad}{spatRaster of solar radiation values for each unit.}
}
\value{
spatRaster with "bio21".
}
\description{
Identifies highest solar radiation unit, potentially using a static index.
}
\keyword{internal}
