% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bios_terra.R
\name{bio22_terra}
\alias{bio22_terra}
\title{bio22_terra: Lowest Solar Radiation Unit}
\usage{
bio22_terra(srad, low_rad_unit = NULL)
}
\arguments{
\item{srad}{spatRaster of solar radiation values for each unit.}
}
\value{
spatRaster with "bio22".
}
\description{
Identifies lowest solar radiation unit, potentially using a static index.
}
\keyword{internal}
