% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bios_fast.R
\name{bio23_fast}
\alias{bio23_fast}
\title{bio23_fast: Solar Radiation Seasonality (CV)}
\usage{
bio23_fast(srad, n_units, cell)
}
\arguments{
\item{srad}{Matrix containing solar radiation values for each unit.}

\item{n_units}{Integer. The total number of temporal units.}

\item{cell}{Vector of original cell IDs.}
}
\value{
Matrix with "bio23", "cell".
}
\description{
Calculates coefficient of variation in solar radiation across units.
}
\keyword{internal}
