% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bios_terra.R
\name{bio23_terra}
\alias{bio23_terra}
\title{bio23_terra: Solar Radiation Seasonality (CV)}
\usage{
bio23_terra(srad)
}
\arguments{
\item{srad}{Matrix containing solar radiation values for each unit.}
}
\value{
spatRaster with "bio23".
}
\description{
Calculates coefficient of variation in solar radiation across units.
}
\keyword{internal}
