% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bios_terra.R
\name{bio24_terra}
\alias{bio24_terra}
\title{bio24_terra: Solar Radiation of Wettest Period}
\usage{
bio24_terra(prad, wettest_period)
}
\arguments{
\item{prad}{spatRaster of solar radiation period means.}

\item{wettest_period}{Vector indicating the index (1-based) of the wettest period.}
}
\value{
spatRaster with "bio24".
}
\description{
Calculates solar radiation mean of the period with the highest precipitation sum.
}
\keyword{internal}
