% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bios_terra.R
\name{bio25_terra}
\alias{bio25_terra}
\title{bio25_terra: Solar Radiation of Driest Period}
\usage{
bio25_terra(prad, driest_period)
}
\arguments{
\item{prad}{spatRaster of solar radiation period means.}

\item{driest_period}{Vector indicating the index (1-based) of the driest period.}
}
\value{
spatRaster with "bio25".
}
\description{
Calculates solar radiation mean of the period with the highest precipitation sum.
}
\keyword{internal}
