% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bios_terra.R
\name{bio26_terra}
\alias{bio26_terra}
\title{bio26_terra: Solar Radiation of Warmest Period}
\usage{
bio26_terra(prad, warmest_period)
}
\arguments{
\item{prad}{spatRaster of solar radiation period means.}

\item{warmest_period}{Vector indicating the index (1-based) of the warmest period.}
}
\value{
spatRaster with "bio26".
}
\description{
Calculates solar radiation mean of the period with the highest temperature mean.
}
\keyword{internal}
