% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bios_fast.R
\name{bio27_fast}
\alias{bio27_fast}
\title{bio27_fast: Solar Radiation of Coldest Period}
\usage{
bio27_fast(speriod, tperiod_min_idx, cell)
}
\arguments{
\item{speriod}{Matrix of solar radiation period means (output from `var_periods`).}

\item{tperiod_min_idx}{Vector indicating the index (1-based) of the coldest period.}

\item{cell}{Vector of original cell IDs.}
}
\value{
Matrix with "bio27", "cell".
}
\description{
Calculates solar radiation mean of the period with the lowest temperature mean.
}
\keyword{internal}
