% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bios_terra.R
\name{bio27_terra}
\alias{bio27_terra}
\title{bio27_terra: Solar Radiation of Coldest Period}
\usage{
bio27_terra(prad, coldest_period)
}
\arguments{
\item{prad}{spatRaster of solar radiation period means.}

\item{coldest_period}{Vector indicating the index (1-based) of the coldest period.}
}
\value{
spatRaster with "bio27".
}
\description{
Calculates solar radiation mean of the period with the lowest temperature mean.
}
\keyword{internal}
