% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bios_fast.R
\name{bio28_fast}
\alias{bio28_fast}
\title{bio28_fast: Mean Moisture of Units}
\usage{
bio28_fast(mois, cell)
}
\arguments{
\item{mois}{Matrix of average moisture for each unit.}

\item{cell}{Vector of original cell IDs.}
}
\value{
Matrix with columns: "bio28", "cell".
}
\description{
Calculates mean moisture across all temporal units.
}
\keyword{internal}
