% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bios_terra.R
\name{bio28_terra}
\alias{bio28_terra}
\title{bio28_terra: Mean Moisture of Units}
\usage{
bio28_terra(mois)
}
\arguments{
\item{mois}{spatRaster of average moisture for each unit.}
}
\value{
spatRaster with "bio28".
}
\description{
Calculates mean moisture across all temporal units.
}
\keyword{internal}
