% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bios_terra.R
\name{bio29_terra}
\alias{bio29_terra}
\title{bio29_terra: Highest Moisture Unit}
\usage{
bio29_terra(mois, high_mois_unit = NULL)
}
\arguments{
\item{mois}{spatRaster of moisture values for each unit.}
}
\value{
spatRaster with "bio29".
}
\description{
Identifies highest moisture unit, potentially using a static index.
}
\keyword{internal}
