% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bios_terra.R
\name{bio30_terra}
\alias{bio30_terra}
\title{bio30_terra: Lowest Moisture Unit}
\usage{
bio30_terra(mois, low_mois_unit = NULL)
}
\arguments{
\item{mois}{spatRaster of moisture values for each unit.}
}
\value{
spatRaster with "bio30".
}
\description{
Identifies lowest moisture unit, potentially using a static index.
}
\keyword{internal}
