% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bios_terra.R
\name{bio31_terra}
\alias{bio31_terra}
\title{bio31_terra: Moisture Seasonality (Standard Deviation)}
\usage{
bio31_terra(mois)
}
\arguments{
\item{mois}{Matrix containing moisture values for each unit.}
}
\value{
spatRaster with "bio31".
}
\description{
Calculates coefficient of variation in moisture across units.
}
\keyword{internal}
