% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bios_fast.R
\name{bio32_fast}
\alias{bio32_fast}
\title{bio32_fast: Moisture of the Most Moist Period}
\usage{
bio32_fast(speriod, speriod_max_idx, cell)
}
\arguments{
\item{speriod}{Matrix of moisture period means (output from `var_periods`).}

\item{speriod_max_idx}{Vector indicating the index (1-based) of the most moist period.}

\item{cell}{Vector of original cell IDs.}
}
\value{
Matrix with "bio32", "cell".
}
\description{
Calculates moisture mean of the most moist period.
}
\keyword{internal}
