% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bios_terra.R
\name{bio32_terra}
\alias{bio32_terra}
\title{bio32_terra: Moisture of the Most Moist Period}
\usage{
bio32_terra(pmois, high_mois_period)
}
\arguments{
\item{pmois}{spatRaster of moisture period means.}

\item{high_mois_period}{Vector indicating the index (1-based) of the most moist period.}
}
\value{
spatRaster with "bio32".
}
\description{
Calculates moisture mean of the most moist period.
}
\keyword{internal}
