% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bios_terra.R
\name{bio33_terra}
\alias{bio33_terra}
\title{bio33_terra: Moisture of the Least Moist Period}
\usage{
bio33_terra(pmois, low_mois_period)
}
\arguments{
\item{pmois}{spatRaster of moisture period means.}

\item{low_mois_period}{Vector indicating the index (1-based) of the least moist period.}
}
\value{
spatRaster with "bio33".
}
\description{
Calculates moisture mean of the least moist period.
}
\keyword{internal}
