% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bios_terra.R
\name{bio34_terra}
\alias{bio34_terra}
\title{bio34_terra: Moisture of Warmest Period}
\usage{
bio34_terra(pmois, warmest_period)
}
\arguments{
\item{pmois}{spatRaster of moisture period means.}

\item{warmest_period}{Vector indicating the index (1-based) of the warmest period.}
}
\value{
spatRaster with "bio34".
}
\description{
Calculates moisture mean of the period with the highest temperature mean.
}
\keyword{internal}
