% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bios_fast.R
\name{bio35_fast}
\alias{bio35_fast}
\title{bio35_fast: Moisture of Coldest Period}
\usage{
bio35_fast(speriod, tperiod_min_idx, cell)
}
\arguments{
\item{speriod}{Matrix of moisture period means (output from `var_periods`).}

\item{tperiod_min_idx}{Vector indicating the index (1-based) of the coldest period.}

\item{cell}{Vector of original cell IDs.}
}
\value{
Matrix with "bio35", "cell".
}
\description{
Calculates moisture mean of the period with the lowest temperature mean.
}
\keyword{internal}
