% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bios_terra.R
\name{bio35_terra}
\alias{bio35_terra}
\title{bio35_terra: Moisture of Coldest Period}
\usage{
bio35_terra(pmois, coldest_period)
}
\arguments{
\item{pmois}{spatRaster of moisture period means.}

\item{coldest_period}{Vector indicating the index (1-based) of the coldest period.}
}
\value{
spatRaster with "bio35".
}
\description{
Calculates moisture mean of the period with the lowest temperature mean.
}
\keyword{internal}
