% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll_fast.R
\name{roll_fast}
\alias{roll_fast}
\title{Tiled, Out-of-Core Rolling Time Series Averaging (Internal) - CORRECTED}
\usage{
roll_fast(
  paths,
  window_size,
  freq,
  step,
  fun,
  output_names_list,
  user_region,
  tile_degrees,
  output_dir
)
}
\description{
Tiled, Out-of-Core Rolling Time Series Averaging (Internal) - CORRECTED
}
\keyword{internal}
