% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_layers.R
\name{write_layers}
\alias{write_layers}
\title{Export Bioclimatic Variables to GeoTIFF}
\usage{
write_layers(
  biovardir,
  save_dir = "bioclimatic",
  clean_temporary_files = FALSE
)
}
\arguments{
\item{biovardir}{Character string. Path to directory with intermediate `.qs` files and `template_info.qs`}

\item{save_dir}{Character string. Path for the final GeoTIFF files. Default: "bioclimatic".}

\item{clean_temporary_files}{Logical. Delete `biovardir` after writing? Default: FALSE.}
}
\value{
None. Writes GeoTIFF files to `save_dir`.
}
\description{
Assembles intermediate `.qs` files into a full vector in memory, performs
cell ID mapping if necessary, and writes final GeoTIFF rasters.
}
\author{
Luis Osorio-Olvera, Gonzalo E. Pinilla-Buitrago
}
